refactor exporter files
